Update factor -> factor_sampler
Replace sample() -> __call__()
