--- keras_cv/core/factor/uniform_factor_test.py
@@ -1,29 +1,0 @@
-# Copyright 2022 The KerasCV Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import tensorflow as tf
-
-from keras_cv import core
-
-
-class UniformFactorSamplerTest(tf.test.TestCase):
-    def test_sample(self):
-        factor = core.UniformFactorSampler(0.3, 0.6)
-        self.assertTrue(0.3 <= factor.sample() <= 0.6)
-
-    def test_config(self):
-        factor = core.UniformFactorSampler(0.3, 0.6)
-        config = factor.get_config()
-        self.assertEqual(config["lower"], 0.3)
-        self.assertEqual(config["upper"], 0.6)
