+++ keras_cv/core/factor_sampler/__init__.py
@@ -11,3 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from keras_cv.core.factor_sampler.constant_factor_sampler import ConstantFactorSampler
+from keras_cv.core.factor_sampler.factor_sampler import FactorSampler
+from keras_cv.core.factor_sampler.uniform_factor_sampler import UniformFactorSampler
