+++ keras_cv/core/factor_sampler/constant_factor_sampler_test.py
@@ -14,15 +14,15 @@
 
 import tensorflow as tf
 
-import keras_cv
+from keras_cv import core
 
 
 class ConstantFactorSamplerTest(tf.test.TestCase):
     def test_sample(self):
-        factor = keras_cv.ConstantFactorSampler(0.3)
+        factor = core.ConstantFactorSampler(0.3)
         self.assertEqual(factor(), 0.3)
 
     def test_config(self):
-        factor = keras_cv.ConstantFactorSampler(0.3)
+        factor = core.ConstantFactorSampler(0.3)
         config = factor.get_config()
         self.assertEqual(config["value"], 0.3)
