+++ keras_cv/core/factor_sampler/factor_sampler.py
@@ -20,15 +20,15 @@ class FactorSampler:
     """FactorSampler represents a strength factor for use in an augmentation layer.
 
     FactorSampler should be subclassed and implement a `__call__()` method that returns
-    a tf.float32, or a float.  This method will be used by preprocessing layers to
-    determine the strength of their augmentation.  The specific range of values
-    supported may vary by layer, but for most layers is the range [0, 1].
+    a tf.float32.  This method will be used by preprocessing layers to determine the
+    strength of their augmentation.  The specific range of values supported may vary by
+    layer, but for most layers is the range [0, 1].
     """
 
-    def __call__(self, shape=None, dtype=tf.float32):
+    def __call__():
         raise NotImplementedError(
             "FactorSampler subclasses must implement a `__call__()` method."
         )
 
-    def get_config(self):
+    def get_config():
         return {}
